Moved projections in separate file
