+++ autoattack/fab_tf.py
@@ -16,6 +16,9 @@ import time
 
 import tensorflow as tf
 
+from autoattack.fab_projections import projection_linf, projection_l2,\
+    projection_l1
+
 #from advertorch.utils import replicate_input
 
 #from .base import Attack
@@ -95,217 +98,6 @@ class FABAttack():
         df, dg = self.model.get_grad_diff_logits_target(imgs, la, la_target)
         
         return df, dg
-        
-    
-    def projection_linf(self, points_to_project, w_hyperplane, b_hyperplane):
-        t = points_to_project.clone()
-        w = w_hyperplane.clone()
-        b = b_hyperplane.clone()
-
-        ind2 = ((w * t).sum(1) - b < 0).nonzero().squeeze()
-        ind2 = self.check_shape(ind2)
-        w[ind2] *= -1
-        b[ind2] *= -1
-
-        c5 = (w < 0).float()
-        a = torch.ones(t.shape).to(self.device)
-        d = (a * c5 - t) * (w != 0).float()
-        a -= a * (1 - c5)
-
-        p = torch.ones(t.shape).to(self.device) * c5 - t * (2 * c5 - 1)
-        indp = torch.argsort(p, dim=1)
-
-        b = b - (w * t).sum(1)
-        b0 = (w * d).sum(1)
-        b1 = b0.clone()
-
-        counter = 0
-        indp2 = indp.unsqueeze(-1).flip(dims=(1, 2)).squeeze()
-        u = torch.arange(0, w.shape[0])
-        ws = w[u.unsqueeze(1), indp2]
-        bs2 = - ws * d[u.unsqueeze(1), indp2]
-
-        s = torch.cumsum(ws.abs(), dim=1)
-        sb = torch.cumsum(bs2, dim=1) + b0.unsqueeze(1)
-
-        c = b - b1 > 0
-        b2 = sb[u, -1] - s[u, -1] * p[u, indp[u, 0]]
-        c_l = (b - b2 > 0).nonzero().squeeze()
-        c2 = ((b - b1 > 0) * (b - b2 <= 0)).nonzero().squeeze()
-        c_l = self.check_shape(c_l)
-        c2 = self.check_shape(c2)
-
-        lb = torch.zeros(c2.shape[0])
-        ub = torch.ones(c2.shape[0]) * (w.shape[1] - 1)
-        nitermax = torch.ceil(torch.log2(torch.tensor(w.shape[1]).float()))
-        counter2 = torch.zeros(lb.shape).long()
-
-        while counter < nitermax:
-            counter4 = torch.floor((lb + ub) / 2)
-            counter2 = counter4.long()
-            indcurr = indp[c2, -counter2 - 1]
-            b2 = sb[c2, counter2] - s[c2, counter2] * p[c2, indcurr]
-            c = b[c2] - b2 > 0
-            ind3 = c.nonzero().squeeze()
-            ind32 = (~c).nonzero().squeeze()
-            ind3 = self.check_shape(ind3)
-            ind32 = self.check_shape(ind32)
-            lb[ind3] = counter4[ind3]
-            ub[ind32] = counter4[ind32]
-            counter += 1
-
-        lb = lb.long()
-        counter2 = 0
-
-        if c_l.nelement != 0:
-            lmbd_opt = (torch.max((b[c_l] - sb[c_l, -1]) / (-s[c_l, -1]),
-                                  torch.zeros(sb[c_l, -1].shape)
-                                  .to(self.device))).unsqueeze(-1)
-            d[c_l] = (2 * a[c_l] - 1) * lmbd_opt
-
-        lmbd_opt = (torch.max((b[c2] - sb[c2, lb]) / (-s[c2, lb]),
-                              torch.zeros(sb[c2, lb].shape)
-                              .to(self.device))).unsqueeze(-1)
-        d[c2] = torch.min(lmbd_opt, d[c2]) * c5[c2]\
-            + torch.max(-lmbd_opt, d[c2]) * (1 - c5[c2])
-
-        return d * (w != 0).float()
-
-    def projection_l2(self, points_to_project, w_hyperplane, b_hyperplane):
-        t = points_to_project.clone()
-        w = w_hyperplane.clone()
-        b = b_hyperplane.clone()
-
-        c = (w * t).sum(1) - b
-        ind2 = (c < 0).nonzero().squeeze()
-        ind2 = self.check_shape(ind2)
-        w[ind2] *= -1
-        c[ind2] *= -1
-
-        u = torch.arange(0, w.shape[0]).unsqueeze(1)
-
-        r = torch.max(t / w, (t - 1) / w)
-        u2 = torch.ones(r.shape).to(self.device)
-        r = torch.min(r, 1e12 * u2)
-        r = torch.max(r, -1e12 * u2)
-        r[w.abs() < 1e-8] = 1e12
-        r[r == -1e12] = -r[r == -1e12]
-        rs, indr = torch.sort(r, dim=1)
-        rs2 = torch.cat((rs[:, 1:],
-                        torch.zeros(rs.shape[0], 1).to(self.device)), 1)
-        rs[rs == 1e12] = 0
-        rs2[rs2 == 1e12] = 0
-
-        w3 = w ** 2
-        w3s = w3[u, indr]
-        w5 = w3s.sum(dim=1, keepdim=True)
-        ws = w5 - torch.cumsum(w3s, dim=1)
-        d = -(r * w).clone()
-        d = d * (w.abs() > 1e-8).float()
-        s = torch.cat(((-w5.squeeze() * rs[:, 0]).unsqueeze(1),
-                      torch.cumsum((-rs2 + rs) * ws, dim=1) -
-                      w5 * rs[:, 0].unsqueeze(-1)), 1)
-
-        c4 = (s[:, 0] + c < 0)
-        c3 = ((d * w).sum(dim=1) + c > 0)
-        c6 = c4.nonzero().squeeze()
-        c2 = ((1 - c4.float()) * (1 - c3.float())).nonzero().squeeze()
-        c6 = self.check_shape(c6)
-        c2 = self.check_shape(c2)
-
-        counter = 0
-        lb = torch.zeros(c2.shape[0])
-        ub = torch.ones(c2.shape[0]) * (w.shape[1] - 1)
-        nitermax = torch.ceil(torch.log2(torch.tensor(w.shape[1]).float()))
-        counter2 = torch.zeros(lb.shape).long()
-
-        while counter < nitermax:
-            counter4 = torch.floor((lb + ub) / 2)
-            counter2 = counter4.long()
-            c3 = s[c2, counter2] + c[c2] > 0
-            ind3 = c3.nonzero().squeeze()
-            ind32 = (~c3).nonzero().squeeze()
-            ind3 = self.check_shape(ind3)
-            ind32 = self.check_shape(ind32)
-            lb[ind3] = counter4[ind3]
-            ub[ind32] = counter4[ind32]
-            counter += 1
-
-        lb = lb.long()
-        alpha = torch.zeros([1])
-
-        if c6.nelement() != 0:
-            alpha = c[c6] / w5[c6].squeeze(-1)
-            d[c6] = -alpha.unsqueeze(-1) * w[c6]
-
-        if c2.nelement() != 0:
-            alpha = (s[c2, lb] + c[c2]) / ws[c2, lb] + rs[c2, lb]
-            if torch.sum(ws[c2, lb] == 0) > 0:
-                ind = (ws[c2, lb] == 0).nonzero().squeeze().long()
-                ind = self.check_shape(ind)
-                alpha[ind] = 0
-            c5 = (alpha.unsqueeze(-1) > r[c2]).float()
-            d[c2] = d[c2] * c5 - alpha.unsqueeze(-1) * w[c2] * (1 - c5)
-
-        return d * (w.abs() > 1e-8).float()
-
-    def projection_l1(self, points_to_project, w_hyperplane, b_hyperplane):
-        t = points_to_project.clone()
-        w = w_hyperplane.clone()
-        b = b_hyperplane.clone()
-
-        c = (w * t).sum(1) - b
-        ind2 = (c < 0).nonzero().squeeze()
-        ind2 = self.check_shape(ind2)
-        w[ind2] *= -1
-        c[ind2] *= -1
-
-        r = torch.max(1 / w, -1 / w)
-        r = torch.min(r, 1e12 * torch.ones(r.shape).to(self.device))
-        rs, indr = torch.sort(r, dim=1)
-        _, indr_rev = torch.sort(indr)
-
-        u = torch.arange(0, w.shape[0]).unsqueeze(1)
-        u2 = torch.arange(0, w.shape[1]).repeat(w.shape[0], 1)
-        c6 = (w < 0).float()
-        d = (-t + c6) * (w != 0).float()
-        d2 = torch.min(-w * t, w * (1 - t))
-        ds = d2[u, indr]
-        ds2 = torch.cat((c.unsqueeze(-1), ds), 1)
-        s = torch.cumsum(ds2, dim=1)
-
-        c4 = s[:, -1] < 0
-        c2 = c4.nonzero().squeeze(-1)
-        c2 = self.check_shape(c2)
-
-        counter = 0
-        lb = torch.zeros(c2.shape[0])
-        ub = torch.ones(c2.shape[0]) * (s.shape[1])
-        nitermax = torch.ceil(torch.log2(torch.tensor(s.shape[1]).float()))
-        counter2 = torch.zeros(lb.shape).long()
-
-        while counter < nitermax:
-            counter4 = torch.floor((lb + ub) / 2)
-            counter2 = counter4.long()
-            c3 = s[c2, counter2] > 0
-            ind3 = c3.nonzero().squeeze()
-            ind32 = (~c3).nonzero().squeeze()
-            ind3 = self.check_shape(ind3)
-            ind32 = self.check_shape(ind32)
-            lb[ind3] = counter4[ind3]
-            ub[ind32] = counter4[ind32]
-            counter += 1
-
-        lb2 = lb.long()
-
-        if c2.nelement() != 0:
-            alpha = -s[c2, lb2] / w[c2, indr[c2, lb2]]
-            c5 = u2[c2].float() < lb.unsqueeze(-1).float()
-            u3 = c5[u[:c5.shape[0]], indr_rev[c2]]
-            d[c2] = d[c2] * u3.float().to(self.device)
-            d[c2, indr[c2, lb2]] = alpha
-
-        return d * (w.abs() > 1e-8).float()
 
     def attack_single_run(self, x, y=None, use_rand_start=False):
         """
@@ -409,17 +201,17 @@ class FABAttack():
                     w = dg2.reshape([bs, -1])
 
                     if self.norm == 'Linf':
-                        d3 = self.projection_linf(
+                        d3 = projection_linf(
                             torch.cat((x1.reshape([bs, -1]), x0), 0),
                             torch.cat((w, w), 0),
                             torch.cat((b, b), 0))
                     elif self.norm == 'L2':
-                        d3 = self.projection_l2(
+                        d3 = projection_l2(
                             torch.cat((x1.reshape([bs, -1]), x0), 0),
                             torch.cat((w, w), 0),
                             torch.cat((b, b), 0))
                     elif self.norm == 'L1':
-                        d3 = self.projection_l1(
+                        d3 = projection_l1(
                             torch.cat((x1.reshape([bs, -1]), x0), 0),
                             torch.cat((w, w), 0),
                             torch.cat((b, b), 0))
@@ -602,17 +394,17 @@ class FABAttack():
                     w = dg2.reshape([bs, -1])
 
                     if self.norm == 'Linf':
-                        d3 = self.projection_linf(
+                        d3 = projection_linf(
                             torch.cat((x1.reshape([bs, -1]), x0), 0),
                             torch.cat((w, w), 0),
                             torch.cat((b, b), 0))
                     elif self.norm == 'L2':
-                        d3 = self.projection_l2(
+                        d3 = projection_l2(
                             torch.cat((x1.reshape([bs, -1]), x0), 0),
                             torch.cat((w, w), 0),
                             torch.cat((b, b), 0))
                     elif self.norm == 'L1':
-                        d3 = self.projection_l1(
+                        d3 = projection_l1(
                             torch.cat((x1.reshape([bs, -1]), x0), 0),
                             torch.cat((w, w), 0),
                             torch.cat((b, b), 0))
